+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.0.10 - in progress
+- Adding new functionality for getting max value of data type (and optionally gene/protein)
 
 ## 0.0.9 - 2021-05-25
 - Updating doc tests, adding new query type